Remove PRId64 from example so it compiles under C89

+++ examples/test_json.c
@@ -104,7 +104,7 @@ static void process_value(json_value* value, int depth)
                         process_array(value, depth+1);
                         break;
                 case json_integer:
-                        printf("int: %10" PRId64 "\n", value->u.integer);
+                        printf("int: %10ld\n", (long)value->u.integer);
                         break;
                 case json_double:
                         printf("double: %f\n", value->u.dbl);
